feat: more chaotic, but less annoying types?
