+++ src/components/layercake-components/Lines.svg.svelte
@@ -1,16 +1,8 @@
 <script lang="ts">
   import { getContext } from 'svelte';
-  import {
-    line,
-    curveCardinal,
-    curveLinear,
-    curveStep,
-    curveStepAfter,
-    curveStepBefore,
-    curveMonotoneX,
-    type CurveFactory
-  } from 'd3-shape';
+  import { line, curveCardinal, type CurveFactory } from 'd3-shape';
   import { type LayerCakeContextType, type LayerCakeGroupedDataGroupValuesType } from '../../lib/types';
+  import { curveMap } from '../../lib/curves';
 
   interface Props {
     curve?: CurveFactory;
@@ -18,15 +10,6 @@
   const { data, xGet, yGet, zGet } = getContext<LayerCakeContextType>('LayerCake');
 
   let { curve }: Props = $props();
-
-  const curveMap = {
-    linear: curveLinear,
-    cardinal: curveCardinal,
-    step: curveStep,
-    stepAfter: curveStepAfter,
-    stepBefore: curveStepBefore,
-    monotoneX: curveMonotoneX
-  };
 
   const renderedLines = $derived(
     $data.flatMap(({ values, config }) => {
@@ -47,7 +30,7 @@
           id: config.id,
           d: line<LayerCakeGroupedDataGroupValuesType>($xGet, $yGet).curve(
             curve || (config.curveType && curveMap[config.curveType]) || curveCardinal
-          ),
+          )(vals),
           stroke: config.colour || (vals[0] ? $zGet(vals[0]) : '#000'),
           dasharray: config.dasharray
         }
